Added simple CSS and fixed size to canvas
I have designed some simple CSS which I will work with for the following weeks, it is still in early development and needs many improvements and research.
I believe that a fixed width and height (which I have chosen 800x1200px) should be placed in the JavaScript code. This will allow us to save the image as a file of a fixed size and it will also make posting drawings easier. This size may change.